\name{colors}
\alias{c.colors}
\title{colors used by plgraphics}
\description{
  Vectors of color names to be used, mainly for distinguishing groups
}
\usage{
c.colors
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  none
}
\value{vector of color names
}

\author{Werner A. Stahel, ETH Zurich}
\examples{
c.colors
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{manip}
