\name{months}
\alias{c.months}
\alias{c.mon}
\alias{c.weekdays}
\alias{c.wdy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{strings for month and weekday names}
\description{
  Vectors of month and weekday names
}
\usage{
c.months
c.mon
c.weekdays
c.wkd
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  none
}
\value{character vector.\cr
  \code{c.months} contains the 12 month names.\cr
  \code{c.mon} same, abbreviated to 3 characters,\cr
  \code{c.weekdays} names of the 7 weekdays\cr
  \code{c.wkd} same, abbreviated to 3 characters,\cr
}

\author{Werner A. Stahel, ETH Zurich}
\examples{
c.weekdays[1:5]
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{manip}
