\name{gentimeaxis}
\alias{gentimeaxis}

\title{Generate attributes for a variable expressing time
}
\description{code{gentimeaxis} generates a variable that has suitable
  attributes for plotting a time
}
\usage{
gentimeaxis(date = NULL, year = NULL, month = NULL, day = NULL,
  hour = NULL, data = NULL, ploptions=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{date}{vector of class \code{Date} or the name of such a
    variable contained in \code{data}
  } 
  \item{year, month, day, hour}{numeric vector giving the
    year, month, day, hour -- or the name of such a
    variable contained in \code{data}
  }
  \item{data}{data.frame, where variables can be found
  }
  \item{ploptions}{list pl options, generated by \code{pl.control}
  }
}
\details{
  This function ... !!! is not yet ready
}
\value{
  Numeric vector giving the suitable coordinate values for plotting and
  the necessary attributes. It contains
  \item{years}{if there are no different months given,}
  \item{days}{(julian) if different days or months are given}
  The result has the attributes
  \item{ticksat}{vector where tickmarks are shown.
  It contains its own attribute \code{small} if secondary ticks are
  suitable.} 
  \item{ticklabels}{May be years, quarters, month names, days, ...}
  \item{ticklabelsat}{vecor of coordinates to place the ticklabels}
  \item{label}{equals "", since the time scale makes it clear enough
    that the axis represents time.}
}
\author{Werner A. Stahel
}
\note{
  \code{gentimeaxis} converts the date into
  \code{julian(date, origin = as.Date("2000-01-01"))}, storing
  these values as the attribute \code{attr(.,"numvalues")}
  (rather than using the default origin at 1970).
  
  This function currently works for time scales from years to hours.
  Minutes and seconds remain to be implemented.
}

\seealso{\code{\link{genvarattributes}}, \code{\link{axis.Date}}
}
\examples{
## call gentimeaxis without 'real' data
gentimeaxis(year=rep(2010:2015, each=12), month=rep(1:12, 6))
gentimeaxis(year=rep(2010:2015, each=12), month=rep(1:12, 6))

## ... derived from data
data(d.river)
d.river$time <- gentimeaxis(date="date", hour="hour", data=d.river)
plyx(O2~time, data=d.river[months(d.river$date)!="September",])
plyx(O2~time, data=d.river[months(d.river$date)!="September",])
plyx(O2~time, data=d.river[1:1000,])
}
\keyword{utilities}

