\name{pltimeaxis}
\alias{pltimeaxis}

\title{Generate attributes for a variable expressing time
}
\description{code{pltimeaxis} generates a variable that has suitable
  attributes for plotting a time
}
\usage{
pltimeaxis(date = NULL, year = NULL, month = NULL, day = NULL,
  hour = NULL, data = NULL, ploptions=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{date}{vector of class \code{Date} or the name of such a
    variable contained in \code{data}
  } 
  \item{year, month, day, hour}{numeric vector giving the
    year, month, day, hour -- or the name of such a
    variable contained in \code{data}
  }
  \item{data}{data.frame, where variables can be found
  }
  \item{ploptions}{list pl options, generated by \code{pl.control}
  }
}
\details{
  This function ... !!! is not yet ready
}
\value{
  Numeric vector giving the suitable coordinate values for plotting and
  the necessary attributes. It contains
  \item{years}{if there are no different months given,}
  \item{days}{(julian) if different days or months are given}
  The result has the attributes
  \item{ticksat}{vector where tickmarks are shown.
  It contains its own attribute \code{small} if secondary ticks are
  suitable.} 
  \item{ticklabels}{May be years, quarters, month names, days, ...}
  \item{ticklabelsat}{vecor of coordinates to place the ticklabels}
  \item{label}{equals "", since the time scale makes it clear enough
    that the axis represents time.}
}
\author{Werner A. Stahel
}
\note{
This function is under development
}

\seealso{\code{\link{genvarattributes}}
}
\examples{
## not ready!
}
\keyword{utilities}

