\name{gentime}
\alias{gentime}
\alias{gentimeaxis}

\title{Generate a variable expressing time or its attributes for plotting
}
\description{\code{gentime} generates a date variable and is an
  extension of \code{\link{as.POSIXct}}.\cr
  \code{gentimeaxis} generates suitable attributes for plotting such a time
}
\usage{
gentime(date = NULL, year = 2000, month = 1, mon = 0, day = 1, mday = 0,
  hour = 0, min = 0, sec = 0, data = NULL, dateformat = "\%Y-\%m-\%d")

gentimeaxis(date = NULL, year = 2000, month = 1, mon = 0, day = 1, mday = 0,
  hour = 0, min = 0, sec = 0, data = NULL, dateformat = "\%Y-\%m-\%d",
  ploptions=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{date}{vector of class \code{POSIXt} or vector to be converted
    into such an object. May also be the name of such a variable
    contained in \code{data}.
    If \code{date} has class \code{POSIXt}, the arguments
    \code{year, month} and \code{day} are ignored. 
  } 
  \item{year, month, day, hour, min, sec}{numeric vectors giving the
    year, month, day of month, hour, minute, second -- or the name of such a
    variable contained in \code{data}.
    \code{day, hour, min, sec} can be fractional, see Details.
    If these arguments are used, the supersede the respective parts in
    \code{date}.
  }
  \item{mon, mday}{numeric vectors giving the month and the day of
    month minus 1, as in the POSIXt format.
    (If \code{month} or \code{day} is also given, it supersedes \code{mon}
    or \code{mday}.)
  }
  \item{data}{data.frame, where variables can be found
  }
  \item{dateformat}{format for \code{date} in case that the latter is
    a character vector
  }
  \item{ploptions}{list pl options, generated by \code{pl.control}
  }
}
\details{
  If \code{hour} is fractional, e.g., 6.2, the fraction is respected,
  that is, it will be the same as time \code{06:12}.
  If \code{min} is also given, the fraction of \code{hour} is ignored.
  Similar for \code{day} and \code{min}.
  \cr
  If \code{hour} is \code{>=24}, the \code{day} is augmented by
  \code{hour\%/\%24} and the hour is set to \code{hour\%\%24}.
  Similar for \code{min} and \code{sec}.
}  
\value{
  For \code{gentime}, a vector of times in \code{POSIXct} format.\cr
  For \code{gentimeaxis}, this is augmented by the attribute
  \item{numvalues}{numerical values used for plotting.
    If years, months or days vary in the data, the units are days.
    Otherwise, they are hours, minutes, or seconds, depending on the
    highest category that varies.
  }
  Unless the dates only cover one of the categories
  (only years differ, or only months, ...), the following plotting
  attributes are added:
  \item{ticksat}{vector where tickmarks are shown.
  It contains its own attribute \code{small} if secondary ticks are
  suitable.} 
  \item{ticklabels}{May be years, quarters, month names, days, ...}
  \item{ticklabelsat}{vecor of coordinates to place the ticklabels}
  \item{label}{equals "", since the time scale makes it clear enough
    that the axis represents time.}
}
\author{Werner A. Stahel
}
\note{
  \code{gentimeaxis} converts the date into
  \code{julian(date, origin = as.Date("2000-01-01"))}, storing
  these values as the attribute \code{attr(.,"numvalues")}
  (rather than using the default origin at 1970).
  
  This function currently works for time scales from years to hours.
  Minutes and seconds remain to be implemented.
}

\seealso{\code{\link{genvarattributes}}, \code{\link{axis.Date}}
}
\examples{
## call gentimeaxis without 'real' data
tt <- gentime(year=rep(2010:2012, each=12), month=rep(1:12, 3))
ta <- gentimeaxis(tt)
str(unclass(ta)) ## str(ta) does not show the attributes

## ... derived from data
data(d.river)
d.river$time <- gentimeaxis(date="date", hour="hour", data=d.river)
plyx(O2~time, data=d.river, subset=months(date)!="September")
plyx(O2~time, data=d.river[months(d.river$date)!="September",])
plyx(O2~time, data=d.river[1:1000,])
}
\keyword{utilities}

