\name{pl.envir}
\title{Environment Containing Information for Generating Pl Graphics}
\alias{pl.envir}
\docType{data}
\description{The environment \code{pl.envir} contains options and
  information generated by the user or the high level pl functions
  to guide many aspects of pl graphical displays.
}
\usage{data("pl.envir")}
%%\format{
%%  The format is:
%%<environment: 0x55c36f0b6ba0> 
%%}
\details{
  The environment contains the following objects:
  \describe{
  \item{usr.ploptions}{The ploptions that are the cuurent
    user choice. It is initiated with the package's
    choices in \code{\link{default.ploptions}} and
    modified whenever ploptions are set
    (and \code{assign} is \code{TRUE}),
    see \code{\link{ploptions}}.
  }
  \item{ploptions}{The last (or current) list used
    (and often modified) by a high level pl function.
  }
  \item{plfeatures, marpar}{Additional features used by pl functions
  }
  \item{pldata}{The last (or current) data set used by pl functions
  }
  \item{pl.scaleFunctions}{scale function known to plgraphics,
    useful for scales to be used for showing axes.}
 }
}
\examples{
pl.envir$usr.ploptions$margin.line   ## same as ...
ploptions("margin.line")

plyx(1:4)
ploptions(margin.line=c(3,1))
ploptions("margin.line")
pl.envir$usr.ploptions$margin.line   
pl.envir$ploptions$margin.line  ## the choice for
  ## the last called high level pl function
pl.envir$plfeatures
pl.envir$marpar  ## used internally by that function

}
\keyword{internal}
