\name{plgraphics-internal}
\alias{i.def}
%% \alias{i.smooth}
%
\alias{is.formula}
\alias{\%nin\%}
%% \alias{print.mregr}
\alias{i.main}
\alias{u.true}
\alias{u.notfalse}
\alias{u.debug}
\alias{DB}
%% ---
\title{Internal regr0 objects}
\description{
  Internal functions of package \pkg{plgraphics} that are exported, but
  should normally not be called by the user directly.
}
\usage{
is.formula(object)
x \%nin\% y
DB(on=TRUE)
}
\arguments{
  \item{x,y, object}{\R objects}
  \item{...}{...}
}
\keyword{internal}
