\name{plpoints}
\alias{plpoints}
\alias{pllines}
\alias{plbars}
\alias{plsmooth}
\alias{plsmoothline}
\alias{plrefline}

\title{Low level plotting functions for the 'pl' system
}
\description{
  These functions do the plotting of points and enhancing lines
  based on the 'pl' paradigm.
}
\usage{
plpoints(x=NULL, y=NULL, type="p", plab=NULL, pch=NULL, col=NULL, lcol=col,
  lty=NULL, lwd=NULL, psize=NULL,
  plargs = NULL, ploptions = NULL, setpar = TRUE, ...)

pllines(x, y, type="l", ...)

plbars(x = NULL, y = NULL, midpointwidth = NULL, 
  plargs = NULL, ploptions = NULL, setpar = TRUE)

plrefline(refline, x=NULL, innerrange=NULL, y=NULL,
  cutrange = c(x = TRUE, y = FALSE), plargs=NULL, ploptions=NULL, ...)

plsmooth(x, y, ysec = NULL, band=NULL, power = NULL,
  plargs = NULL, ploptions = NULL)
plsmoothline(smoothline, x, y, ysec = NULL,
  plargs=NULL, ploptions = NULL, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x, y}{coordinates for the horizontal and veritical axis,
    respectively. If \code{NULL}, they will be retrieved from
    the first and second column of \code{plargs$pldata}.
  }
  \item{plab}{labels for displaying points. Overrides labels provided by
    \code{plargs$pdata[["plab"]]}.
  }
  \item{col}{color points}
  \item{lcol}{color for lines}
  \item{type}{type of displaying points. See \code{?\link{points}}.}
  \item{pch, psize, lty, lwd}{... and \code{col} in \code{plpoints}:
    plotting character(s), size and color of plotting points,
    and line type. %% to be used for \code{plpoints}.
    Overrides other settings, defined in \code{plargs}.
  }
  \item{midpointwidth}{for \code{plbars}: determines the length of the
    segments that mark the midpoint. See Details.
  }
  \item{refline}{for \code{plrefline}:
    A vector giving intercept and slope of a straight line, 
    or a function that returns this list, such as \code{lm}}
  \item{innerrange}{for \code{plrefline}: inner range in x direction -
    only needed of the refline should be clipped at a range different
    from the \code{innerrange} attribute of the horizontal variable}
  \item{cutrange}{for \code{plrefline}: logical vector of length 2:
    should the reference line(s) be cut at the inner plotting ranges in
    x- and y-direction?
    Otherwise, it will be continued outside it with the appropriate
    transformation.
  }
  \item{smoothline}{for \code{plsmoothline}: result of a smooth fitting}
  \item{ysec}{for \code{plsmooth, plsmoothline}:
    matrix of secondary y values. The smooths generated or given by its
    columns will be drawn thinner and with paled color.
  }
  \item{band}{logical: should a band (e.g., a confidence band)
    be drawn together with the smooth?
  }
  \item{power}{for \code{plsmooth}: smooth will be calcutated for
    \code{y^power} and the back-transformed. Usually, \code{power=0}.
  }
  \item{plargs, ploptions}{result of \code{\link{pl.control}}, see Details
  }
  \item{setpar}{logical: should graphical parameters
    \code{cex, mar, mgp} be set according to \code{ploptions}?
    See Details.
  }    
  \item{\dots}{absorbs extra arguments}
}
\value{
  \code{plsmooth} invisibly returns the data.frame needed for
  drawing the smooth line. The other functions return \code{NULL}
}
\details{
  For \code{plpoints}, the first arguments, \code{x} and \code{y}
  can be formulas, and an argument \code{data} can be given.
  These arguments then have the same meaning as in \code{\link{plyx}}.

  For \code{plbars}, the argument \code{midpoints} determines the
  length of the segments that mark the midpoint relative to the default,
  which is proportional to the range of the plotting area and inversely
  proportional to the number of (finite) observations.

  \code{plargs} and \code{ploptions} may be specified explicitly,
  but they are usually generated by calling \code{pl.control}.

  The argument \code{setpar} is used for setting the graphical
  parameters \code{cex, mar, mgp} according to \code{ploptions}?
  This is needed if the high level pl function has changed \code{mar},
  since this change has been reversed when the function was left.
  By default, these graphical parameters will be retieved from
  \code{.plargs$ploptions}.

}
  
\author{Werner A. Stahel
}

\seealso{\link{pl.control}
}
\examples{
plyx(Sepal.Width ~ Sepal.Length, data=iris)

## again, each step separately
t.dt <- .plargs$pldata
.plargs$grid <- TRUE
plframe(t.dt$Sepal.Length, t.dt$Sepal.Width, plargs=.plargs)
plsmooth(t.dt$Sepal.Length, t.dt$Sepal.Width, plargs=.plargs)
t.plab <- plmark(t.dt$Sepal.Length, t.dt$Sepal.Width, markextremes=0.03,
                 plargs=.plargs)
plpoints(t.dt$Sepal.Length, t.dt$Sepal.Width, plargs=.plargs, plab=t.plab) 

## plbars

data(d.river)
dd <- plsubset(d.river, 1:2000)
da <- aggregate(dd[,3:7], dd[,"date",drop=FALSE], mean, na.rm=TRUE)
ds <- aggregate(dd[,3:7], dd[,"date",drop=FALSE], sd, na.rm=TRUE)
plyx(O2~date, data=da, type="n")
td <- da$O2 + outer(ds$O2, c(0,-1,1))
plbars(y = td, midpointwidth=0.1)

}


\keyword{aplot}

